Extract `replica_set/shard_transfer.rs`
